Fix sign-compare warning when compiling with GCC 14 (#1749)
* Fix sign-compare warning when compiling with GCC 14
* Use C size_t instead of C++ std::size_t

+++ src/opentimelineio/stackAlgorithm.cpp
@@ -144,7 +144,7 @@ _normalize_tracks_lengths(
         }
     }
 
-    for (int i = 0; i < tracks.size(); i++)
+    for (size_t i = 0; i < tracks.size(); i++)
     {
         Track*       old_track      = tracks[i];
         RationalTime track_duration = old_track->duration(error_status);
